[CORE] - Basic project setup
