+++ pubspec.yaml
@@ -32,6 +32,12 @@ dependencies:
     sdk: flutter
 
   cupertino_icons: ^1.0.8
+  flutter_riverpod: ^3.3.1
+  dio: ^5.9.2
+  flutter_secure_storage: ^10.0.0
+  google_sign_in: ^7.2.0
+  sign_in_with_apple: ^7.0.1
+  font_awesome_flutter: ^11.0.0
 
 dev_dependencies:
   flutter_test:
